create_dataset_funnel/requirements.txt
imutils==0.5.4
matplotlib==3.6.0
numpy==1.17.4
opencv_python_headless==4.6.0.66
Pillow==9.3.0
rembg==2.0.25
scipy==1.7.3
sympy==1.11.1
matplotlib==3.6.0
